requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Pose Estimation
mediapipe>=0.10.0
opencv-python>=4.8.0

# Data & Config
pyyaml>=6.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Development
pytest>=7.3.0
black>=23.0.0
